# backend/docs/api/phone-store-openapi.yaml
openapi: 3.0.3
info:
  title: Phone Store E-Commerce Platform API
  version: 1.0.0
  description: API documentation for the Phone Store E-Commerce Platform backend services.
servers:
  - url: http://localhost:8080
    description: local dev server
...
paths:
...
  /api/v1/admin/categories:
    post:
      security:
        - BearerAuth: []
      tags:
        - Category Admin
      summary: API Endpoint
      description: API Description
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                SlugExists:
                  summary: API Endpoint
                  value:
                    errorCode: CATEGORY_SLUG_EXISTS
                    message: Error
                    correlationId: db6ef274-129b-4e08-9fd1-cb4be6c1e345
        '422':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      security:
        - BearerAuth: []
      tags:
        - Category Admin
      summary: API Endpoint
      description: API Description
      operationId: adminListCategories
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
        - name: keyword
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: API Description
        '403':
          description: API Description
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/categories/{categoryId}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Category Admin
      summary: API Endpoint
      operationId: getCategoryById
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      security:
        - BearerAuth: []
      tags:
        - Category Admin
      summary: API Endpoint
      description: API Description
      operationId: updateCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: API Description
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: API Description
        '422':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/categories/{categoryId}/status:
    patch:
      security:
        - BearerAuth: []
      tags:
        - Category Admin
      summary: API Endpoint
      description: API Description
      operationId: changeCategoryStatus
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: API Description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/categories/tree:
    get:
      tags:
        - Category
      summary: API Endpoint
      description: API Description
      operationId: getActiveCategoryTree
      responses:
        '200':
          description: API Description
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/cart:
    get:
      tags:
        - Cart
      summary: Get cart info
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/cart/items:
    post:
      tags:
        - Cart
      summary: Add to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Cart
      summary: Update cart item quantity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/cart/items/{itemId}:
    delete:
      tags:
        - Cart
      summary: Delete Cart Item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/cart/merge:
    post:
      tags:
        - Cart
      summary: Merge Guest Cart
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    BadRequest:
      description: L?i d? li?u d?u vào (Thi?u ho?c sai d?nh d?ng tham s?)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            MissingParameter:
              summary: Thi?u tham s?
              value:
                errorCode: BAD_REQUEST
                message: Tham s? b?t bu?c không du?c d? tr?ng
                correlationId: e3b0c442-uuid-abc
            InvalidFormat:
              summary: Sai d?nh d?ng UUID/Email
              value:
                errorCode: BAD_REQUEST
                message: Ð?nh d?ng email không h?p l?
                correlationId: e3b0c442-uuid-def
    Unauthorized:
      description: K? m?o danh (Chua dang nh?p ho?c Token h?t h?n)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            TokenExpired:
              summary: JWT h?t h?n
              value:
                errorCode: UNAUTHORIZED
                message: Access token dã h?t h?n
                correlationId: e3b0c442-uuid-123
    Forbidden:
      description: B? t? ch?i (Không d? quy?n h?n truy c?p chuyên sâu)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            AccessDenied:
              summary: Vu?t quy?n
              value:
                errorCode: FORBIDDEN
                message: Tài kho?n c?a b?n không có d?c quy?n (Admin/Manager) d? g?i API này
                correlationId: e3b0c442-uuid-789
    NotFound:
      description: Không tìm th?y (Tài nguyên không t?n t?i trên DB)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            ResourceNotFound:
              summary: D? li?u tr?ng
              value:
                errorCode: RESOURCE_NOT_FOUND
                message: Không tìm th?y d? li?u id=9999 (Entity Not Found)
                correlationId: e3b0c442-uuid-444
    UnprocessableEntity:
      description: L?i logic nghi?p v? (Ki?m tra constraint/business th?t b?i)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            BusinessConflict:
              summary: H?t t?n kho
              value:
                errorCode: BUSINESS_CONFLICT
                message: S? lu?ng t?n kho không d? d? th?c hi?n yêu c?u!
                correlationId: e3b0c442-uuid-555
            ValidationFailed:
              summary: Vi ph?m constraint
              value:
                errorCode: VALIDATION_FAILED
                message: Validation Error
                fieldErrors:
                  quantity: S? lu?ng ph?i l?n hon 0
                  price: Giá bán không du?c âm
                correlationId: e3b0c442-uuid-666
    InternalServerError:
      description: L?i máy ch? c?c b? (Exception backend không mong mu?n)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            ServerError:
              summary: Null Pointer / Database Down
              value:
                errorCode: INTERNAL_SERVER_ERROR
                message: Ðã x?y ra l?i h? th?ng nghiêm tr?ng, vui lòng liên h? Admin
                correlationId: e3b0c442-uuid-999
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required:
        - errorCode
        - message
        - correlationId
      properties:
        errorCode:
          type: string
          description: API Description
          example: BAD_REQUEST
        message:
          type: string
          description: API Description
          example: Invalid username or password
        correlationId:
          type: string
          format: uuid
          description: API Description
          example: a6b7c8d9-e0f1-42a3-b4c5-d6e7f8a9b0c1
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          description: API Description
          example:
            username: Username is required
            password: Password is required
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 150
          example: example
        parentId:
          type: string
          format: uuid
          nullable: true
          example: a6b7c8d9-e0f1-42a3-b4c5-d6e7f8a9b0c2
        description:
          type: string
          example: example
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
          default: ACTIVE
          example: ACTIVE
        sortOrder:
          type: integer
          minimum: 0
          default: 0
          example: 0
    CategoryResponse:
      type: object
      required:
        - id
        - name
        - slug
        - status
        - sortOrder
        - subCategories
      properties:
        id:
          type: string
          format: uuid
          example: a6b7c8d9-e0f1-42a3-b4c5-d6e7f8a9b0c2
        parentId:
          type: string
          format: uuid
          nullable: true
          example: null
        name:
          type: string
          example: example
        slug:
          type: string
          example: dien-thoai
        description:
          type: string
          nullable: true
          example: example
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
          example: ACTIVE
        sortOrder:
          type: integer
          example: 0
        subCategories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        createdAt:
          type: string
          format: date-time
          example: '2026-07-16T15:00:00Z'
...